Activate vox before other xontribs
When vox was activated after apipenv, then opening a new shell in a
folder with a Pipfile resulted in calls to vox before the xontrib had
been loaded. This resulted in improper error messages about missing
the vox program and meant the Pipenv was not auto-loading in that
folder.

By moving activation of vox before these other xontribs, those xontribs
can make use of vox internally